Separate special classes into own module
